Move state related functionality into state itself
